ReConstruct mechanism of rotating Cube
I hope that this structure would be the last one. The point concept of
this structure is to separate rotation and translation. In this case,
rotation means rotating itself, not rotated around origin.

When the cube is rotated, one of R,L,U,D,F,B operations, each cubePiece actually moving toward
another position while rotating itself.

Bue this structure also has a problem about rotating.
When a lot of rotation is done, cubePiece does not rotate as we want..