Add style in mobile web view
The problem is that cube length is dynamically changed. because of that,
we need to optimize cube position and length whenever display width
changed or just force user to use one type of display only. Also, I need
to figure out the width at first loading because of cubeLength value